+++ cpp/temperature-sensor-2/main/sleeping/sleeping.h
@@ -23,5 +23,5 @@
 using time_point = std::chrono::time_point<std::chrono::system_clock>;
 
 int DeepSleep(time_point soft_sleep_tp, time_point hard_sleep_tp,
-              std::uint32_t temperature_threshold);
+              std::int16_t temperature_threshold);
 #endif  // SLEEPING_SLEEPING_H_
